requirements.txt

feedparser
requests
beautifulsoup4
jinja2
python-dotenv
PyYAML
python-dateutil
